refactor: Use ArtifactToken with the artifact cache

+++ plugins/org.eclipse.osee.ats.core.client/src/org/eclipse/osee/ats/core/client/util/AtsChangeSet.java
@@ -119,7 +119,7 @@ public class AtsChangeSet extends AbstractAtsChangeSet {
 
          List<Artifact> changedArts = new LinkedList<>();
          for (Long changedUuid : results.getChangedWorkitemUuids()) {
-            Artifact artifact = ArtifactCache.getActive(changedUuid.intValue(), AtsUtilCore.getAtsBranch());
+            Artifact artifact = ArtifactCache.getActive(changedUuid, AtsUtilCore.getAtsBranch());
             if (artifact != null) {
                changedArts.add(artifact);
             }
